Passing 404 test for #12 (responds with not found when comment_id doesn't exist)

+++ models/reviews.models.js
@@ -130,11 +130,21 @@ exports.insertComment = (newComment, review_id) => {
 };
 
 exports.deleteComment = (removedComment) => {
-  return db
-    .query(`DELETE FROM comments WHERE comment_id = $1`, [removedComment])
-    .then((comment) => {
-      console.log(comment.rows, "<--- comment.rows");
-      // Returning an empty array?
-      return comment.rows;
+  if (removedComment) {
+    let checkCommentsStr = `SELECT * FROM comments WHERE comment_id = $1`;
+    return db.query(checkCommentsStr, [removedComment]).then((res) => {
+      if (!res.rows.length) {
+        return Promise.reject({
+          status: 404,
+          message: "Not found: The comment_id does not exist",
+        });
+      } else {
+        return db
+          .query(`DELETE FROM comments WHERE comment_id = $1`, [removedComment])
+          .then((comment) => {
+            return comment.rows;
+          });
+      }
     });
+  }
 };
